+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Next
+
+- **[Feature]** Expose `epsilons` as a public field.
+
 # 0.1.1 (2019-03-30)
 
 - **[Feature]** Derive additional traits: `Copy`, `Clone`, `PartialOrd`, `Ord`, `Hash` ([#1](https://github.com/open-flash/rust-swf-fixed/issues/1))